requirements.txt
fastapi
uvicorn
python-multipart
insightface
onnxruntime
opencv-python
numpy
Pillow
pymongo
motor
python-dotenv
pydantic
pydantic-settings
google-cloud-storage
python-multipart
pytz